moved implementation setting to output_model base class
