+++ bhmm/output_models/gaussian.py
@@ -22,15 +22,6 @@ class GaussianOutputModel(OutputModel):
     """
 
 
-    # implementation codes
-    __IMPL_PYTHON__ = 0
-    __IMPL_C__ = 1
-
-    # implementation used
-    __impl__= __IMPL_PYTHON__
-
-
-
     def __init__(self, nstates, means=None, sigmas=None):
         """
         Create a 1D Gaussian output model.
@@ -69,26 +60,6 @@ class GaussianOutputModel(OutputModel):
             self.sigmas = np.zeros([nstates], dtype=dtype)
 
         return
-
-
-    def set_implementation(self, impl):
-        """
-        Sets the implementation of this module
-
-        Parameters
-        ----------
-        impl : str
-            One of ["python", "c"]
-
-        """
-        if impl.lower() == 'python':
-            self.__impl__ = self.__IMPL_PYTHON__
-        elif impl.lower() == 'c':
-            self.__impl__ = self.__IMPL_C__
-        else:
-            import warnings
-            warnings.warn('Implementation '+impl+' is not known. Using the fallback python implementation.')
-            self.__impl__ = self.__IMPL_PYTHON__
 
 
     def __repr__(self):
